Allow to provide virtio-net to a VM when we don't measure performance there
If we assign a device to a VM along with virtual IOMMU, the VM can in
turn assign the device to its own VM. This would make the VM not having
any network device.  Therefore we provide virtio-net to the VM along
with an assigned device unless we measure performance there.

Signed-off-by: Jintack Lim <jintack@cs.columbia.edu>